swap Overture engine to SedonaDB; rename geo crate → transport
- replace duckdb workspace dep with sedona (git, pinned to
  apache-sedona-db-0.4.0) + matching arrow 57.1; features ["aws","geo"]
  (geos dropped — needs a system GEOS not present here)
- rename crate geo → transport, avoiding a clash with georust's `geo`
  that sedona pulls in transitively
- add overture module: opens a SedonaContext (smoke-tested)
- derive per-(device, UTC day) bboxes in SQL (GROUP BY + MIN/MAX)
  rather than a hand-rolled Rust fold

Builds need CXX=/usr/bin/clang (the blocked clang++ workaround).
