+++ apps/lookout/Cargo.toml
@@ -22,12 +22,20 @@ async-trait = "0.1"
 clap = { version = "4", features = ["derive"] }
 # `bundled` compiles a pinned SQLite from source, so the recorder doesn't depend on
 # a system libsqlite3 being present or new enough. It also builds SQLite with the
-# R*Tree module (`-DSQLITE_ENABLE_RTREE`), used by the geo enrichment for a spatial
-# index over the `transport` table — no separate rusqlite feature is needed for it.
+# R*Tree module (`-DSQLITE_ENABLE_RTREE`), used by the transport enrichment for a
+# spatial index over the `transport` table — no separate rusqlite feature is needed.
 rusqlite = { version = "0.40", features = ["bundled"] }
-# `bundled` compiles libduckdb from source (no system libduckdb needed); the geo
-# `enrich` bin uses it in-process to fetch Overture transport data from S3.
-duckdb = { version = "1", features = ["bundled"] }
+# SedonaDB: a Rust-native, DataFusion/GeoArrow spatial engine the transport `enrich`
+# bin uses in-process to read Overture transport GeoParquet from S3. Not published to
+# crates.io, so pinned to a release tag. We select `aws` (S3 object store) + the
+# pure-Rust `geo` geometry backend, and drop the default `geos`/`tg` backends: `geos`
+# links a *system* GEOS library that isn't present here (enable it later once GEOS is
+# installed). The build still needs a C++ toolchain for transitive C/C++ crates — the
+# Justfile sets CXX=/usr/bin/clang to work around the blocked clang++ shim.
+sedona = { git = "https://github.com/apache/sedona-db", tag = "apache-sedona-db-0.4.0", default-features = false, features = ["aws", "geo"] }
+# Must match the arrow version SedonaDB pins, so `RecordBatch`es returned by queries
+# line up with the arrow types we decode them with.
+arrow = "57.1"
 md5 = "0.7"
 tempfile = "3"
 shared = { path = "crates/shared" }
