+++ apps/lookout/crates/transport/src/lib.rs
@@ -1,0 +1,10 @@
+//! Transport enrichment: reads a lookout SQLite archive and derives the spatial
+//! extent of its GPS data, to fetch Overture transport data intersecting that extent.
+//!
+//!   - [`groups`] — group fixes by `(device_id, UTC day)` and reduce each to a bbox.
+//!   - [`archive`] — read the `gps` table those groups are built from.
+//!   - [`overture`] — query Overture transport GeoParquet via SedonaDB.
+
+pub mod archive;
+pub mod groups;
+pub mod overture;
